refactor(v1): move tunnels into the v1.interception.tunnel subpackage
Split the single tunnel module into a package mirroring runtimes/: base.py
(the Tunnel contract), prime.py / modal.py / custom.py (the implementations),
and __init__.py with the make_tunnel(config) factory + _tunnel_cls dispatch.
Tunnels are built from a config via make_tunnel (no method on the config), the
host-side counterpart to make_runtime. Also tighten the EnvConfig.interception
docstring ("from a remote runtime"). No behavior change.
