+++ verifiers/v1/interception/__init__.py
@@ -23,6 +23,7 @@ from verifiers.v1.interception.tunnel import (
     ModalTunnel,
     PrimeTunnel,
     Tunnel,
+    make_tunnel,
 )
 
 __all__ = [
@@ -40,4 +41,5 @@ __all__ = [
     "PrimeTunnel",
     "ModalTunnel",
     "CustomTunnel",
+    "make_tunnel",
 ]
