--- verifiers/v1/interception/tunnel.py
@@ -1,115 +1,0 @@
-"""Tunnels: the contract for making the host interception server reachable from a remote harness.
-
-The interception server runs on the host; a harness in a remote runtime reaches it over a tunnel.
-`Tunnel` is that contract — how to build the reachable URL, and what address/port the server must
-bind. Each `InterceptionConfig` owns the implementation matching its type (`config.tunnel()`):
-`PrimeTunnel` (prime_tunnel/frpc), `ModalTunnel` (`modal.forward`), `CustomTunnel` (a bring-your-own
-reverse proxy). The host-side counterpart to a `Runtime`: a `Runtime.expose` publishes a port
-*inside* a sandbox; a `Tunnel.expose` publishes a *host* port outward.
-"""
-
-import contextlib
-from abc import ABC, abstractmethod
-from collections.abc import AsyncIterator
-from typing import TYPE_CHECKING, ClassVar
-
-from verifiers.v1.runtimes.base import host_endpoint, open_tunnel
-
-if TYPE_CHECKING:
-    from verifiers.v1.interception.config import (
-        BaseInterceptionConfig,
-        UrlInterceptionConfig,
-    )
-
-
-class Tunnel(ABC):
-    """Makes the host interception server reachable from a remote harness runtime. One per
-    interception type, owned by its config (`config.tunnel()`); lightweight and stateless beyond
-    the config it holds."""
-
-    bind_host: ClassVar[str] = "127.0.0.1"
-    """Address the interception server must bind for this tunnel to reach it. Loopback by default —
-    frpc / a BYO proxy connect over localhost on the same host."""
-
-    single_server: ClassVar[bool] = False
-    """Whether this is a single fixed endpoint shared by every rollout (so the pool never grows past
-    one server). True only for `CustomTunnel` — one BYO URL is structurally one server."""
-
-    def __init__(self, config: "BaseInterceptionConfig") -> None:
-        self.config = config
-
-    @property
-    def bind_port(self) -> int:
-        """Fixed local port the interception server must bind (0 = an ephemeral port)."""
-        return 0
-
-    @contextlib.asynccontextmanager
-    async def reachable(self, port: int, *, is_local: bool) -> AsyncIterator[str]:
-        """Yield a URL the harness's runtime uses to reach the host interception server on `port`. A
-        local runtime shares the host network → localhost (no tunnel, whatever the type); a remote
-        one is bridged by `expose`."""
-        if is_local:
-            yield f"http://127.0.0.1:{port}"
-        else:
-            async with self.expose(port) as url:
-                yield url
-
-    @abstractmethod
-    def expose(self, port: int) -> contextlib.AbstractAsyncContextManager[str]:
-        """An async context manager yielding a public URL that reaches the host's `port` from a
-        remote runtime. Entered only when the harness runtime is remote; torn down on exit."""
-
-
-class PrimeTunnel(Tunnel):
-    """Expose the host port via prime_tunnel (frpc) — the shared host endpoint. The default; works
-    from any host with prime credentials."""
-
-    @contextlib.asynccontextmanager
-    async def expose(self, port: int) -> AsyncIterator[str]:
-        async with host_endpoint(port, is_local=False) as url:
-            yield url
-
-
-class ModalTunnel(Tunnel):
-    """Expose the host port via Modal's own forwarding (`modal.forward`). Requires the framework to
-    run inside a Modal container — `modal.forward` raises `InvalidError` otherwise. Binds `0.0.0.0`:
-    Modal forwards to the container's routable interface, not its loopback, so a loopback-only server
-    is unreachable. Safe — that server runs inside an isolated container whose only ingress is the
-    tunnel itself."""
-
-    bind_host: ClassVar[str] = "0.0.0.0"
-
-    @contextlib.asynccontextmanager
-    async def expose(self, port: int) -> AsyncIterator[str]:
-        try:
-            import modal
-        except ModuleNotFoundError as e:
-            raise ModuleNotFoundError(
-                "modal interception requires the Modal SDK; install `verifiers[modal]`."
-            ) from e
-        stack = contextlib.AsyncExitStack()
-
-        async def _start() -> str:
-            tunnel = await stack.enter_async_context(modal.forward(port))
-            return str(tunnel.url).rstrip("/")
-
-        try:
-            yield await open_tunnel(_start, f"modal host tunnel (port {port})")
-        finally:
-            await stack.aclose()
-
-
-class CustomTunnel(Tunnel):
-    """Bring your own reverse proxy: open no tunnel, reach the server at the configured `url` (which
-    you front the fixed local `port` with). One URL is one server, shared by every rollout."""
-
-    config: "UrlInterceptionConfig"
-    single_server: ClassVar[bool] = True
-
-    @property
-    def bind_port(self) -> int:
-        return self.config.port
-
-    @contextlib.asynccontextmanager
-    async def expose(self, port: int) -> AsyncIterator[str]:
-        yield self.config.url
